common.utils: automatic refactor: WWW
Move `error_page` and `timestamp` to `www.utils`.
