+++ webserver.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 import common.time
 import common.url
+import www.utils
 from common import utils
 from common.config import config
 from www.server import app
@@ -21,7 +22,7 @@ import www.quotes
 app.secret_key = config["session_secret"]
 app.add_template_filter(common.time.nice_duration)
 app.add_template_filter(utils.ucfirst)
-app.add_template_filter(utils.timestamp)
+app.add_template_filter(www.utils.timestamp)
 app.add_template_filter(common.url.https)
 app.add_template_filter(common.url.noproto)
 app.csrf_token = app.jinja_env.globals["csrf_token"]
